Extract Security Update helper methods
